requirements.txt
astroid==2.4.1
cycler==0.10.0
future==0.18.2
imageio==2.8.0
isort==4.3.21
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pip==9.0.1
progressbar==2.5
pybullet==2.7.7
pylint==2.5.2
pyparsing==2.4.7
pyquaternion==0.9.5
python-dateutil==2.8.1
PyYAML==5.3.1
scikit-learn==0.23.1
scipy==1.4.1
setuptools==39.0.1
six==1.14.0
threadpoolctl==2.0.0
toml==0.10.1
torch
torchvision
tqdm==4.46.0
typed-ast==1.4.1
wrapt==1.12.1
visdom==0.1.8.9
scikit-image==0.17.2